+++ common/elastic/index.go
@@ -1,0 +1,6 @@
+package elastic
+
+type Index struct {
+	Name string
+	Body string
+}
